requirements.txt
flask
flask-smorest
python-dotenv
marshmallow
flask-sqlalchemy
pyopenssl
flask-jwt-extended
argon2-cffi
gunicorn
nltk
pypdf
python-docx
beautifulsoup4
